Added useCart hook
